implemented dining hall logic
