+++ src/main/kotlin/com/madalv/Global.kt
@@ -1,0 +1,14 @@
+package com.madalv
+
+object Cfg {
+    val host = "kitchen"
+    const val timeUnit = 100
+    const val nrTables = 10
+    const val nrWaiters = 4
+    const val maxItemsPerOrder = 10
+    const val orderIdMax = 500
+
+    const val maxTableWait: Long = 30
+    const val waitTimeCoeff = 1.3
+}
+
