--- src/main/kotlin/com/madalv/GlobalConfig.kt
@@ -1,14 +1,0 @@
-package com.madalv
-
-object Cfg {
-    const val host: String = "kitchen"
-    const val timeUnit = 1000
-    const val nrTables = 10
-    const val nrWaiterst = 4
-    const val maxItemsPerOrder = 10
-    const val orderIdMax = 500
-
-    const val maxTableWait: Long = 30
-    const val waitTimeCoeff = 1.3
-}
-
