+++ src/main/kotlin/com/madalv/Menu.kt
@@ -1,0 +1,4 @@
+package com.madalv
+
+class Menu {
+}
