+++ src/main/kotlin/com/madalv/Table.kt
