--- src/main/kotlin/com/madalv/TableState.kt
@@ -1,7 +1,0 @@
-package com.madalv
-
-enum class TableState {
-    FREE,
-    WAITING_TO_ORDER,
-    WAITING_FOR_ORDER
-}
