+++ src/main/kotlin/com/madalv/Waiter.kt
@@ -1,0 +1,4 @@
+package com.madalv
+
+class Waiter {
+}
